while updating README, felt like it made sense to move some functions from create_lignin to kmc_functions
